refactor: 패키지 분리
vendingmachine
ㄴcontroller
    Machine.java
ㄴmodel
    ㄴcash
        ㄴdomain
            Coin.java
        ㄴservice
            Cashier.java
    ㄴitem
        ㄴdomain
            Item.java
        ㄴservice
            ItemManager.java
            ItemParser.java
    ㄴshared
        Error.java
        Validator.java
ㄴview
    Display.java
Application.java

MVC계층 - Model의 도메인 영역 - 영역 내 계층 순으로 분리
