+++ src/main/java/vendingmachine/Application.java
@@ -1,5 +1,7 @@
 package vendingmachine;
 
+import vendingmachine.controller.Machine;
+
 public class Application {
 	public static void main(String[] args) {
 		Machine machine = new Machine();
